Add handler for 'ATTR' tokens "{ attr:value }"
Basically the same as 'WORD' and 'REGEXP' tokens, except operates on the
<attr> property of the input token rather than the 'word' property, as
the 'WORD' and 'REGEXP' components do.

Additionally, have the re -> postfix parser store <attr> and <value> as
additional components in the array representing a RE component. For
example:

  'a'           -> [[ 'WORD'    , 'a' ]]
  '/a/'         -> [[ 'REGEXP'  , 'a' ]]
  '{ tag: a }'  -> [[ 'ATTR'    , 'tag', 'a' ]]
  '{ tag: /a/ }'-> [[ 'ATTR'    , 'tag', /a/ ]]

Note that the value associated with an attribute matcher will either be
a string or a RegExp instance, depending on what is parsed, and should
be handled appropriately for each case.